[ATOM-SGL][Attn refrac] Split full attention backend helpers
